+++ projects/ngx-ukis-utilities/src/lib/ngx-ukis-utilities.ts
@@ -1,0 +1,15 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'ngx-ngx-ukis-utilities',
+  imports: [],
+  template: `
+    <p>
+      ngx-ukis-utilities works!
+    </p>
+  `,
+  styles: ``
+})
+export class NgxUkisUtilities {
+
+}