+++ projects/ngx-ukis-utilities/src/public-api.ts
@@ -1,0 +1,7 @@
+/*
+ * Public API Surface of ngx-ukis-utilities
+ */
+
+export * from './lib/ngx-ukis-utilities';
+export * from './lib/pipes';
+export * from './lib/dynamic-component/dynamic-component';